+++ src/lib/components/Drawer/Drawer.tsx
@@ -1,6 +1,7 @@
 import React, { Fragment, ReactNode, useEffect } from 'react';
 import { createPortal } from 'react-dom';
 import { Backdrop } from '../Backdrop';
+import styles from './Drawer.module.scss';
 
 export interface IDrawerProps {
 	children?: ReactNode;
@@ -14,9 +15,9 @@ export const Drawer = (props: IDrawerProps) => {
 	const { children, position = 'left', onClickBackdrop } = props;
 
 	useEffect(() => {
-		document.body.classList.add('drawer-open');
+		document.body.classList.add(styles.drawerOpen);
 		return () => {
-			document.body.classList.remove('drawer-open');
+			document.body.classList.remove(styles.drawerOpen);
 		};
 	}, []);
 
@@ -35,7 +36,7 @@ export const Drawer = (props: IDrawerProps) => {
 
 	return createPortal(
 		<Fragment>
-			<div className="drawer" style={getStyles()}>
+			<div className={styles.drawer} style={getStyles()}>
 				{children}
 			</div>
 			<Backdrop onClick={handleClickBackdrop} />